+++ lib/Controller/EditorController.php
@@ -1392,13 +1392,33 @@ class EditorController extends Controller {
         }
 
         $newFileUri = null;
+        $newFileType = $toExtension;
         $documentService = new DocumentService($this->trans, $this->config);
         $key = $this->fileUtility->getKey($file);
         $fileUrl = $this->getUrl($file, $user);
+        $thumbnail = ['first' => false];
         try {
-            $newFileUri = $documentService->getConvertedUri($fileUrl, $ext, $toExtension, $key);
+            $response = $documentService->sendRequestToConvertService(
+                $fileUrl,
+                $ext,
+                $toExtension,
+                $key,
+                false,
+                null,
+                false,
+                $thumbnail,
+            );
+
+            if (isset($response->error)) {
+                $documentService->processConvServResponceError($response->error);
+            }
+
+            if (isset($response->endConvert) && $response->endConvert === true) {
+                $newFileUri = $response->fileUrl;
+                $newFileType = $response->fileType;
+            }
         } catch (\Exception $e) {
-            $this->logger->error("getConvertedUri: " . $file->getId(), ["exception" => $e]);
+            $this->logger->error("sendRequestToConvertService: " . $file->getId(), ["exception" => $e]);
             return $this->renderError($e->getMessage());
         }
 
@@ -1410,9 +1430,9 @@ class EditorController extends Controller {
         }
 
         $fileNameWithoutExt = substr($fileName, 0, strlen($fileName) - strlen($ext) - 1);
-        $newFileName = $fileNameWithoutExt . "." . $toExtension;
+        $newFileName = "$fileNameWithoutExt.$newFileType";
 
-        $mimeType = $this->config->getMimeType($toExtension);
+        $mimeType = $this->config->getMimeType($newFileType);
 
         return new DataDownloadResponse($newData, $newFileName, $mimeType);
     }