+++ lib/DocumentService.php
@@ -121,10 +121,20 @@ class DocumentService {
      * @param bool $is_async - Perform conversions asynchronously
      * @param string $region - Region
      * @param bool $toForm - Convert to form
+     * @param array $thumbnail - Settings for the thumbnail
      *
      * @return array
      */
-    public function sendRequestToConvertService($document_uri, $from_extension, $to_extension, $document_revision_id, $is_async, $region = null, $toForm = false) {
+    public function sendRequestToConvertService(
+        $document_uri,
+        $from_extension,
+        $to_extension,
+        $document_revision_id,
+        $is_async,
+        $region = null,
+        $toForm = false,
+        $thumbnail = [],
+    ) {
         $documentServerUrl = $this->config->getDocumentServerInternalUrl();
 
         if (empty($documentServerUrl)) {
@@ -167,6 +177,10 @@ class DocumentService {
             $data["pdf"] = [
                 "form" => true
             ];
+        }
+
+        if (!empty($thumbnail)) {
+            $data['thumbnail'] = $thumbnail;
         }
 
         $opts = [