+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##
 ## Changed
+- fix opening a shared link when group access to the app is restricted
 - Nextcloud v30 is no longer supported
 - show warning popup when jwt secret is empty and success message otherwise