+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Changed
 - fix opening a shared link when group access to the app is restricted
 - Nextcloud v30 is no longer supported
+- delete extra permissions on file delete
 
 ## 9.11.0
 ## Changed