+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Nextcloud v30 is no longer supported
 - delete extra permissions on file delete
 
+## Added
+- plugin description and useful links in admin settings
+
 ## 9.11.0
 ## Changed
 - compatible with Nextcloud 32