+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 ## Changed
 - fix opening a shared link when group access to the app is restricted
 - Nextcloud v30 is no longer supported
+- delete extra permissions on file delete
 - replaced general save success message with dedicated messages for each settings section
+
+## Added
+- plugin description and useful links in admin settings
 
 ## 9.11.0
 ## Changed